test_rdt: numerous updates
Allows mapping the labels understood by a decision tree to those of the
test set and also mapping test set labels to aggregated output labels.

A confusion matrix is printed at the end so it's possible to get some
insight into which labels are classified confidently or not.

The code for inference is now a part of test_rdt itself instead of
depending on infer.cc because test_rdt aim to support testing of older
trees and it's useful to be isolated from infer.cc which may evolve
in line with the latest trees that we are using.

The tool will try and recognise common mistakes with testing an older
decision tree with newer training data where the label indices don't
match.