rdt: take lock setting interrupt to trigger thread exit
When we're trying to trigger the worker threads to quit at the end of
training we set 'interrupted = true' and then broadcast the condition
variable that the threads will be waiting on when there's no work.

There was a race between threads checking the interrupted state in a
loop and starting another pthread_cond_wait where interrupted could be
set but then lead to a dead lock.