Add rdt-update tool for adding 'labels' array to .json trees
It's important to be clear about the semantics of each label that a
decision tree can pick so we can make sure we only use a tree with
training data that is compatible (when testing or re-loading a tree
to continue training more levels).

This tool was used to add an array of label names to our pre-existing
decision trees so they will remain consistent with newer decision trees
(which allows us to compare accuracy metrics).