Add native json-to-jnt tool
Actually the plan is to avoid the need to generate .jnt files but an
easily testable stepping stone is to have a native json-to-jnt tool that
can derive the same arrays of joint data from a joint map and the
per-frame .json files. The code can later be swapped with existing code
that loads .jnt files and the json-to-jnt tools removed.

Avoiding the need for .jnt files simplifies our training instructions
but also lets us avoid storing joint map dependent data within a
training directory which could potentially get out of sync with joint
map changes.