rdt: generalize how work is scheduled across threads
This avoids having a special main thread that synchronizes node
processing. The main thread just becomes thread 0 in the pool of
workers.

The work queue entries are now comprised of a callback function and some
arbitrary void *user_data.

There is a schedule_node_work() function that's responsible for adding
new node training jobs to the work queue by taking pending entries from
the training queue.

Nodes are split into 'shards' to parallelize processing a single node
(as before) and there will be a separate job for each shard of a node.

When a shard is finished, we queue a job to process the results of all
shards for that node. This is done speculatively while we don't yet know
whether all other shards for the node are complete. Only one of these
follow-up jobs will do any real work after all shards have completed.

After processing the results of a node (and potentially adding new nodes
to the training queue) then schedule_node_work() is called so we have a
continuous feedback cycle until there's nothing left in the
training_queue and all the threads go idle (implicitly signalling that
we've finished).

The scheduler will try and ensure there's enough work queued in advance
so that worker threads should never be left waiting for more work to do.
We may now be processing more than one node at the same time.