rdt: Only load labels to generate random samples
At the start of training we randomly pick e.g. 2000 pixels in every
labelled training image and those are the pixels that we train the
decision tree with.

We only ever need to know the labels of these selected pixels.

This patch makes sure to free the label images as soon as we've picked
our pixels and associates the label with the pixel instead of repeatedly
looking up the label based on the pixel index and coordinates.

This patch also compresses the size of each pixel from 16 bytes to 8
bytes, by only keeping 16bit xy coordinates, an 8bit label and a 24bit
image index. Previously we had 32bit xy coordinates, a 32bit index and
then a padding word at the end of the struct.

In total for 100k, 172x224 images with 2k samples per image we go from
using 14.76GB to 9.31GB

For 500k images we go from 73.79GB to 46.52GB which is now possible
with 64GB of RAM.