rdt: add debug_post_inference property
This adds a debug option that causes us to write label inference images
for all the root node pixels that we sampled across all images.

Notably the inference is done using the same fixed-point sampling
function that's used while training which makes the images a useful
reference for checking that our sampling for runtime inference is
equivalent.