rdt: add a way to control training a subset of the nodes
So that it's possible to split the training of decision trees up across
multiple machines in a cluster this add batch_divider and batch_number
properties whereby we only train nodes where:

  node.id % batch_divider == batch_number

If the number of nodes for the current depth is lower than the batch
divider then we will also train that node.

For example to split the work across four machines then each instance
would be run with a batch_divider of 4 and then batch numbers of 0, 1, 2
and 3 would be used for each machine.

Some other tool can then be created that will recombine the partial
trees trained like this back into to the final, complete tree.