infer: use most pedantic fixed-point label inference
Currently there are three ifdef guarded version of the label inference
code for the case where we sample with round-nearest UVs and discrete
gradients in millimeters. This enables the version that is closest to
the train_rdt.cc code since the others don't have *exactly* the same
rounding details currently and at least with overfitting test cases we
can see some incorrect pixel classifications as a result.

We can revisit this later when looking at performance and testing with
fully trained trees and larger test sets.